add songs upsert orm

+++ nghe-backend/src/file/property.rs
@@ -2,6 +2,6 @@
 #[cfg_attr(test, derive(PartialEq, Eq, fake::Dummy, Clone, Copy))]
 pub struct File<F> {
     pub hash: u64,
-    pub size: u64,
+    pub size: u32,
     pub format: F,
 }
